Move element card to components
