+++ frontend/src/components/ElementCard.vue
